requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
google-cloud-storage==1.20.0
Pillow==6.2.0
Redis
